Fixed unsigned/signed comparison warning

+++ src/priority-bp/impl/Image.cpp
@@ -246,7 +246,7 @@ void ImageScalable::ScaleDown()
 	wxASSERT(m_depth >= 0);
 
 	// If there's no mask for the next lower depth, create one from the current resolution.
-	if (m_depth == m_resolutions.size() - 1)
+	if (static_cast<unsigned int>(m_depth) == m_resolutions.size() - 1)
 	{
 		const Image& imageToScaleDown = GetCurrentResolution();
 		m_resolutions.push_back(new ImageScaledDown(imageToScaleDown));
